Move tests from tests/ to src/
